+++ examples/advanced/index.html
@@ -4,15 +4,6 @@
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1.0, user-scalable=yes">
   <title>Cycle 😎 Deepstream</title>
-
-  <style>
-   body {
-     margin: 0;
-     font-family: sans-serif;
-     background-color: #f1f1f1;
-   }
-  </style>
-  
 </head>
 <body unresolved>
   <x-app></x-app>
