--- examples/advanced/src/containers/AppDrawer.scss
@@ -1,12 +1,0 @@
-.app-drawer-content .center {
-    text-align: center;
-}
-
-.button.green {
-    background-color: #00897b;
-    color: #fff;
-}
-
-.app-drawer-content .avatar {
-    width: 100px;
-}
