+++ examples/advanced/src/containers/AppDrawer.ts
@@ -3,7 +3,7 @@ import { h, div, br, span, a, button, input } from '@cycle/dom'
 import * as deepstream from '../actions/deepstream'
 import { Sources, Sinks } from '../types'
 
-import './AppDrawer.scss'
+import '../../styles/AppDrawer.scss'
 
 const userImages = {
   john: require('file-loader!../../images/john.jpg'),
@@ -25,6 +25,36 @@ function view(userData = null) {
     }
   }
 
+  const randomGreeting = (name) => {
+    const greetings = [
+      `Ask me why, ${name}?`,
+      `Love me do, ${name}.`,
+      `Stop my mind from wandering, ${name}!`,
+      `A four of fish and finger pies.`,
+      `${name}, don't you know? We all live in a yello submarine!`,
+      `Where do they all belong, ${name}?`,
+      `${name} got the bill and Rita paid it.`,
+      `Do you want to know a secret, ${name}?`,
+      `${name}, let it be, let it be.`,
+      `Alright, alright, ${name}.`,
+      `There will be a show tonight on trampoline.`,
+      `Baby take a chance with me.`,
+      `We'd like to thank you once again.`,
+      `${name}? Oh, he just plays it like that.`,
+      `Send me a postcard, ${name}, drop me a line.`,
+      `Deliver the letter, sooner the better.`,
+      `I'd like to be, under the sea.`,
+      `All these places have their moments.`,
+      `Just to dance with you is everything.`,
+      `Tomorrow may rain, so follow the sun.`,
+      `${name}, don't you know the latest dance?`,
+      `It's the next best thing to be, ${name}, free as a bird.`,
+      `${name}, feeling two foot small. You've got to hide your love away.`,
+      `${name}, I read the news today, oh boy!`
+    ]
+    return greetings[Math.floor(Math.random() * greetings.length)]
+  }
+
   const appDrawerLayout = (<any>document.querySelector('app-drawer-layout'))
   if (userData) {
     // SIDE EFFECT: Turn off forceNarrow mode:
@@ -37,7 +67,10 @@ function view(userData = null) {
         h('paper-button#drawer-sign-in.sign-in.button.green', {
           attrs: { raised: '' },
           hook: hooks
-        }, 'sign out')
+        }, 'sign out'),
+        h('hr'),
+        h('span', randomGreeting(userData.name)),
+        h('hr'),
       ])
     ])
   } else {
