+++ examples/advanced/styles/app.scss
@@ -1,0 +1,14 @@
+body {
+    margin: 0;
+    font-family: sans-serif;
+    background-color: #ECEFF1;
+    color: #355A64;
+}
+
+app-header {
+    background-color: #607D8B !important;
+}
+
+.center {
+    text-align: center;
+}
